Move to types directory + created melees
