+++ src/types/melees/index.ts
@@ -1,0 +1,2 @@
+export type Melees = "Knife"
+    | "Baseball Bat"
